fix: Make `AsyncConnectionPool::close` less opinionated
Rather than acquiring and awaiting all connections upon `close`, we
now simply close the semaphore and close any connections currently in
the pool.

We now document that the onus is on the user to ensure that
all connection handles are closed if they wish to handle all
`Connection::close` results. We reflect this in the `Node::close`
implementation and error.