refactor: Rename `da_block_number` -> `l2_block_number`, `get_state_value` -> `query_state` (#22)
* refactor: Rename get_state_value back to query_state

Just realised I unintentionally changed the name of this in the
`node-db` implementation. I think the original convention with the
`query_` prefix is a better fit as the state can change.

* refactor: Rename da_block_num to l2_block_num

Also clarifies that the L2 block should be the block containing the L2
block containing the DA proof for the contract.