fix: Various small fixes for state derivation and validation streams (#109)
* fix: Various small fixes for state derivation and validation streams

This changes the node handle's `join` implementation to use `try_join!`
internally rather than `select!`. This means that we return early in the
case that any of the streams error, but in the case that a stream closes
successfully we also wait for the other streams to close successfully.

This fixes a bug in both the node and builder CLIs where the validation
stream can end before the state derivation stream, which would result in
closing the node DB early while the state derivation stream is still
using it.

Also fixes mismatched branches in the node CLI for pending future in the
case that no node streams are enabled.

Also fixes the behaviour for state derivation and validation in the case
that no relayer is provided. Rather than returning immediately due to
the block notification channel closing, we re-arrange the stream logic
to first validate or derive state from all available blocks in the DB
from the current progress point, and then wait on the block notification
channel only after we have reached the head of the chain.

* feat: Publish node 0.7 version with stream fixes