feat: Replace `get_block_number` with more general `get_block_header`
Prior to this PR, the only way to acquire the timestamp for a block
given its address was to first get the block number, then list out the
blocks.

This replaces `get_block_number` with a more generally useful
`get_block_header` query fn that returns all header info for a block
(currently just the number and timestamp).