refactor: Move the `node::db` module into `node_db::pool` (#115)
* refactor: Move the `node::db` module into `node_db::pool`

Previously, the node's `ConnectionPool` type was defined in the
`essential-node` crate. This moves the implementation to a feature-gated
`pool` module within the `essential-node-db` crate.

This allows downstream crates to use the node connection pool
abstraction without also needing to include the node stream logic.

* refactor: Take advantage of node ConnectionPool in `essential-relayer`

Previously the node's `ConnectionPool` implementation was not available
to the relayer, as it was in the `essential-node` crate. Now that it has
been moved to the `essential-node-db` crate, we can simplify some parts
of the relayer implementation by taking advantage of the node-specific
`ConnectionPool` type.